# [RFC] Use [GeneratedRegex] in UnicodeCharEscape