--- a/src/Onigwrap/Onigwrap.csproj
+++ b/src/Onigwrap/Onigwrap.csproj
@@ -2,10 +2,9 @@
 
   <PropertyGroup>
     <OutputType>Library</OutputType>
-    <TargetFramework>netstandard2.0</TargetFramework>
+    <TargetFrameworks>net8.0;netstandard2.0</TargetFrameworks>
     <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
     <IsPackable>true</IsPackable>
-    <LangVersion>8.0</LangVersion>
     <Version>$([System.IO.File]::ReadAllText("$(MSBuildProjectDirectory)/../../VERSION"))</Version>
     <Description>Simple wrapper for Oniguruma.</Description>
     <Authors>Aikawa Yataro</Authors>