--- a/src/Onigwrap/UnicodeCharEscape.cs
+++ b/src/Onigwrap/UnicodeCharEscape.cs
@@ -1,20 +1,36 @@
+using System;
 using System.Text.RegularExpressions;
 
 namespace Onigwrap
 {
-    public class UnicodeCharEscape
+    public partial class UnicodeCharEscape
     {
-        private static Regex UNICODE_WITHOUT_BRACES_PATTERN = new Regex("\\\\x[A-Fa-f0-9]{2,8}");
+        private const string WithoutBracesPattern = "\\\\x[A-Fa-f0-9]{2,8}";
+        private static readonly Regex UNICODE_WITHOUT_BRACES_PATTERN =
+
+#if NET8_0_OR_GREATER
+            WithoutBracesRegex();
+
+        [GeneratedRegex(WithoutBracesPattern)]
+        private static partial Regex WithoutBracesRegex();
+#else
+            new Regex(WithoutBracesPattern);
+#endif
 
         public static string AddBracesToUnicodePatterns(string pattern)
         {
             return UNICODE_WITHOUT_BRACES_PATTERN.Replace(pattern, (m) =>
             {
-                string prefix = "\\x";
+                const string prefix = "\\x";
 
-                return string.Concat(
-                    prefix,
-                    "{", m.Value.Substring(prefix.Length), "}");
+                var substr =
+#if NET8_0_OR_GREATER
+                    m.Value.AsSpan(prefix.Length);
+#else
+                    m.Value.Substring(prefix.Length);
+#endif
+
+                return string.Concat(prefix, "{", substr, "}");
             });
         }