I'm not sure how much perf this might give you, but:

I was having a test of a few things, and wondered whether using ```GeneratedRegex``` here to build the regular expressions at compile time would be useful for performance or anything, but that does mean adding a .NET target on top of the .NET Standard one.

I added .NET 8.0, as that might also give the possibility of using [LibraryImport](https://learn.microsoft.com/en-gb/dotnet/standard/native-interop/pinvoke-source-generation) for calling the native oniguruma functions, and enabling all the trimming and AOT analyzers etc.

Any thoughts?