@Numpsy 2024-08-15
There are also some Regexes in TextMateSharp that could perhaps be done in the same way, but I thought I'd start at the bottom of the stack :-)