@danipen 2024-08-22
Hey @Numpsy, thanks for your patience—I was on vacation, so I'm just catching up now. I appreciate your suggestions! Here are my thoughts:

**1. Adding .NET 8 Target:**

- I see the benefits of adding `.NET 8`, like better performance, AOT, and cool features like `GeneratedRegex`. But it also means dealing with conditional compilation, which can make the code harder to maintain and more error-prone. This lib was originally built to be a `.NET Standard 2.0` library for broad compatibility. I'd love to hear more about the potential trade-offs and benefits of introducing `.NET 8`. Could you elaborate a bit more on that? (Apologies, I'm not very experienced with building for multiple targets.)
  - When `.netstandard2` would be used and when `.net 8` would be used? 
  - Will this impact the NuGet package size? 
  - Can users benefit from .NET 8 features just by consuming the NuGet package, or would they need to include the full project? 

**2. Performance Gain from GeneratedRegex:**

- I'm curious to see how much of a performance boost we could get from using `GeneratedRegex`. If it's significant, it might be worth the added complexity of targeting `.NET 8`, and we could apply that across the TextMateSharp repository. Some benchmarks or comparisons would really help us understand the impact in our specific context.

**3. Native AOT and LibraryImport:**

- I agree that if we go full Native AOT, switching from `DllImport` to `LibraryImport` makes sense. This would align nicely with the native interop improvements in .NET 8 and could give us even more performance benefits.

Overall, I'm a bit cautious about adding `.NET 8`. But I'm definitely open to exploring it further if the performance gains are worth it. It might be a good idea to run some tests or benchmarks, and evaluate advantages and disadvantages of adding a new target framework to help us make an informed decision.

Looking forward to your thoughts!