@Numpsy 2024-08-22
For the general multi-targetting points:
> When .netstandard2 would be used and when .net 8 would be used?

.NET 8 would be used by .NET 8/9/Future consumers, everything else would use .NET Standard 2.0

> Will this impact the NuGet package size?

It would contain two versions of onigwrap.dll, though the current dll is 12KB so it's quite trivial compared to the size of all the native dlls.

> Can users benefit from .NET 8 features just by consuming the NuGet package

That's the idea.