@Numpsy 2024-08-22
I don't have a great deal of experience with the generated regexes, but I tried using the unit tests as a base for a benchmarkdotnet project at https://github.com/aikawayataro/Onigwrap/compare/master...Numpsy:Onigwrap:simple_bench and got this:

Current code:
```
BenchmarkDotNet v0.14.0, Windows 10 (10.0.19045.4780/22H2/2022Update) (Hyper-V)
11th Gen Intel Core i7-1185G7 3.00GHz, 1 CPU, 4 logical and 2 physical cores
.NET SDK 9.0.100-preview.7.24407.12
  [Host]     : .NET 8.0.8 (8.0.824.36612), X64 RyuJIT AVX-512F+CD+BW+DQ+VL+VBMI
  DefaultJob : .NET 8.0.8 (8.0.824.36612), X64 RyuJIT AVX-512F+CD+BW+DQ+VL+VBMI


| Method                                                    | Mean     | Error   | StdDev  | Gen0   | Allocated |
|---------------------------------------------------------- |---------:|--------:|--------:|-------:|----------:|
| Unicode_Patterns_Of_Len_2_Without_Branches                | 442.8 ns | 7.63 ns | 6.76 ns | 0.1082 |     680 B |
| Unicode_Patterns_Of_Len_3_Without_Branches                | 443.5 ns | 8.55 ns | 9.85 ns | 0.1082 |     680 B |
| Unicode_Patterns_Of_Len_4_Without_Branches                | 433.8 ns | 6.49 ns | 6.07 ns | 0.1121 |     704 B |
| Several_Unicode_Patterns_Without_Branches                 | 575.8 ns | 6.52 ns | 5.78 ns | 0.1688 |    1064 B |
| Unicode_Patterns_Of_Len_7_Without_Branches                | 698.6 ns | 5.96 ns | 5.28 ns | 0.2394 |    1504 B |
| Unicode_Patterns_Of_Len_8_Without_Branches                | 726.5 ns | 9.81 ns | 8.70 ns | 0.2432 |    1528 B |
| Already_Escaped_Unicode_Chars_Should_Not_Be_Escaped_Again | 170.1 ns | 1.38 ns | 1.15 ns |      - |         - |
```

This PR
```
BenchmarkDotNet v0.14.0, Windows 10 (10.0.19045.4780/22H2/2022Update) (Hyper-V)
11th Gen Intel Core i7-1185G7 3.00GHz, 1 CPU, 4 logical and 2 physical cores
.NET SDK 9.0.100-preview.7.24407.12
  [Host]     : .NET 8.0.8 (8.0.824.36612), X64 RyuJIT AVX-512F+CD+BW+DQ+VL+VBMI
  DefaultJob : .NET 8.0.8 (8.0.824.36612), X64 RyuJIT AVX-512F+CD+BW+DQ+VL+VBMI


| Method                                                    | Mean      | Error    | StdDev   | Gen0   | Allocated |
|---------------------------------------------------------- |----------:|---------:|---------:|-------:|----------:|
| Unicode_Patterns_Of_Len_2_Without_Branches                | 341.71 ns | 5.382 ns | 5.035 ns | 0.0978 |     616 B |
| Unicode_Patterns_Of_Len_3_Without_Branches                | 328.53 ns | 4.558 ns | 3.806 ns | 0.0978 |     616 B |
| Unicode_Patterns_Of_Len_4_Without_Branches                | 324.13 ns | 3.722 ns | 3.299 ns | 0.1016 |     640 B |
| Several_Unicode_Patterns_Without_Branches                 | 395.68 ns | 3.571 ns | 3.165 ns | 0.1540 |     968 B |
| Unicode_Patterns_Of_Len_7_Without_Branches                | 500.86 ns | 7.146 ns | 6.684 ns | 0.2165 |    1360 B |
| Unicode_Patterns_Of_Len_8_Without_Branches                | 499.35 ns | 9.285 ns | 8.685 ns | 0.2203 |    1384 B |
| Already_Escaped_Unicode_Chars_Should_Not_Be_Escaped_Again |  60.92 ns | 0.516 ns | 0.457 ns |      - |         - |
```

Which makes the change seem quite measurable, though this is only a small part of the library functionality.

I'll see if I can get a benchmark of DllImport / LibraryImport, but I might not have time till the weekend