@aikawayataro 2024-08-23
I believe `LibraryImport` will not make much difference. From what I understand from the documentation, it will make marshalling faster and more optimizable for AOT, but the code doesn't marshal stuff this way.
It's good to see an actual benchmark as there could be more things to consider, will see if extra complexity is worthwhile.

As for shipping the dotnet 8 target, I think that's actually a good idea, an extra 12K of compressed space for extra performance, could also ship dotnet 6 this way

BTW, if speed is really an issue, should we profile the code first? I think we better find some hot spots that make a real difference